visual bell: add circle rendering mode
This commit experiments with alternative rendering of visual bell,
and as such it's extensively/excessively configurable.

It renders an overlay of a circle with configurable colors (base,
outline), position and size. Defaults to the center of the window.

Size can be relative to window or chars width, and allows for instance
to place it at the middle/side of a top/bottom tmux status-bar with
exact char height to make it look like a flashing LED at the bar, etc.